langgraph/chatbot_advanced/pyproject.toml
[project]
name = "langgraph-advanced"
version = "0.1.0"
description = "Advanced LangGraph patterns and demos"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community==0.4.1",
    "langchain-openai==1.1.12",
    "python-dotenv==1.2.2",
]
